# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ānvīkṣikī publication platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers:
    get:
      operationId: listPapers
      tags: [papers]
      summary: List papers with filters
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [draft, published] }
        - name: category
          in: query
          schema: { type: string }
        - name: peerReviewed
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Papers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperListResponse"
    post:
      operationId: createPaper
      tags: [papers]
      summary: Create paper (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperInput"
      responses:
        "201":
          description: Created paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
  /papers/latest:
    get:
      operationId: getLatestPapers
      tags: [papers]
      summary: Get latest published papers
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Latest papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paper"
  /papers/{slug}:
    get:
      operationId: getPaper
      tags: [papers]
      summary: Get paper by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
        "404":
          description: Not found
    patch:
      operationId: updatePaper
      tags: [papers]
      summary: Update paper (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperUpdate"
      responses:
        "200":
          description: Updated paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paper"
    delete:
      operationId: deletePaper
      tags: [papers]
      summary: Delete paper (admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /papers/{slug}/download:
    post:
      operationId: incrementPaperDownload
      tags: [papers]
      summary: Increment paper download count
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated download count
          content:
            application/json:
              schema:
                type: object
                properties:
                  downloadCount:
                    type: integer
                required: [downloadCount]

  # ─── CATEGORIES ───────────────────────────────────────────────────────────
...
  /submissions:
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit article or paper proposal (public, rate-limited)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "429":
          description: Rate limited
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List all submissions (admin)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [pending, accepted, rejected] }
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    patch:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update submission status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionUpdate"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # ─── NEWSLETTER ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Paper:
      type: object
      properties:
        id:              { type: integer }
        slug:            { type: string }
        title:           { type: string }
        abstract:        { type: string }
        pdfUrl:          { type: ["string", "null"] }
        categoryId:      { type: integer }
        categorySlug:    { type: string }
        categoryName:    { type: string }
        authorId:        { type: integer }
        authorName:      { type: string }
        status:          { type: string, enum: [draft, published] }
        peerReviewed:    { type: boolean }
        citationText:    { type: ["string", "null"] }
        publicationType: { type: ["string", "null"] }
        publishedAt:     { type: ["string", "null"] }
        downloadCount:   { type: integer }
        tags:            { type: array, items: { type: string } }
      required: [id, slug, title, abstract, categoryId, categorySlug, categoryName, authorId, authorName, status, peerReviewed, downloadCount, tags]
    PaperInput:
      type: object
      properties:
        slug:            { type: string }
        title:           { type: string }
        abstract:        { type: string }
        pdfUrl:          { type: string }
        categoryId:      { type: integer }
        authorId:        { type: integer }
        status:          { type: string, enum: [draft, published] }
        peerReviewed:    { type: boolean }
        citationText:    { type: string }
        publicationType: { type: string }
        tags:            { type: array, items: { type: string } }
      required: [title, abstract, categoryId, authorId]
    PaperUpdate:
      type: object
      properties:
        title:           { type: string }
        abstract:        { type: string }
        pdfUrl:          { type: string }
        categoryId:      { type: integer }
        authorId:        { type: integer }
        status:          { type: string, enum: [draft, published] }
        peerReviewed:    { type: boolean }
        citationText:    { type: string }
        publicationType: { type: string }
        tags:            { type: array, items: { type: string } }
    PaperListResponse:
      type: object
      properties:
        papers:  { type: array, items: { $ref: "#/components/schemas/Paper" } }
        total:   { type: integer }
        page:    { type: integer }
        limit:   { type: integer }
        hasMore: { type: boolean }
      required: [papers, total, page, limit, hasMore]

    # ─── Category ────────────────────────────────────────────────────────
...
    Submission:
      type: object
      properties:
        id:        { type: integer }
        name:      { type: string }
        email:     { type: string }
        type:      { type: string, enum: [article, paper] }
        title:     { type: string }
        abstract:  { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        status:    { type: string, enum: [pending, accepted, rejected] }
        createdAt: { type: string }
      required: [id, name, email, type, title, status, createdAt]
    SubmissionInput:
      type: object
      properties:
        name:     { type: string }
        email:    { type: string }
        type:     { type: string, enum: [article, paper] }
        title:    { type: string }
        abstract: { type: string }
        notes:    { type: string }
      required: [name, email, type, title]
    SubmissionUpdate:
      type: object
      properties:
        status: { type: string, enum: [pending, accepted, rejected] }
      required: [status]

    # ─── Newsletter ──────────────────────────────────────────────────────
...